Add trait data for species we are using
Joe Wright has given us permission to use this data and make available with our code.
Data soruced from BCI_traits_20101220.csv, reducing columns to only one trait and
removing species which are not included in our analysis.

This file is missing from earlier vesrions of history so they will not run without adding it in

closes #52, closes #17