Placeholder code for alternative way of getting data to cluster
Idea will be to submit jobs that run the two commands at the bottom:
`update_data` and `update_models`.  They can be submitted as normal
jobs and will run on either worker (can both be submitted).  They return
the sha of the pulled repo and can be safely rerun.  Once these have
run, then the actual jobs can be submitted.

However, doing this requires opening the repo, as we have no way of
authenticating with github easily here.